FarmTokenContracts/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
web3==6.15.1
python-dotenv==1.0.1
requests==2.31.0
